requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
Deprecated==1.3.1
ecdsa==0.19.1
Flask==3.1.3
Flask-Caching==2.3.1
Flask-Limiter==4.1.1
flask-marshmallow==1.4.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
ordered-set==4.1.0
packaging==26.0
pyasn1==0.6.2
PyJWT==2.11.0
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.1.6
wrapt==2.1.1
